requirements.txt
azure-mgmt-resource
azure-mgmt-network
azure-mgmt-compute
azure-identity
haikunator
Flask
Flask-RESTFul
flask_pymongo
flask-restful-swagger